requirements.txt
pytest
pre-commit
requests
pyzotero
python-dotenv
pymupdf
build
types-requests
numpy
flask
flask-session
gunicorn==20.1.0
langchain
html2text
pypdf
openai
pinecone-client